Moved mesh generation to Preprocessor project.
