+++ ISAAR.MSolve.FEM/Interpolation/InterpolationTri3.cs
@@ -3,8 +3,8 @@ using System.Collections.Generic;
 using System.Text;
 using ISAAR.MSolve.FEM.Entities;
 using ISAAR.MSolve.FEM.Interpolation.Inverse;
-using ISAAR.MSolve.FEM.Meshes;
 using ISAAR.MSolve.Geometry.Coordinates;
+using ISAAR.MSolve.Geometry.Shapes;
 
 // Tri3 nodes:
 // 1
