+++ ISAAR.MSolve.SamplesConsole/FEM/Cantilever2D.cs
@@ -7,13 +7,14 @@ using ISAAR.MSolve.Analyzers;
 using ISAAR.MSolve.Discretization.Interfaces;
 using ISAAR.MSolve.FEM.Elements;
 using ISAAR.MSolve.FEM.Entities;
-using ISAAR.MSolve.FEM.Meshes;
-using ISAAR.MSolve.FEM.Meshes.Custom;
-using ISAAR.MSolve.FEM.Meshes.GMSH;
+using ISAAR.MSolve.Geometry.Shapes;
 using ISAAR.MSolve.Logging.VTK;
 using ISAAR.MSolve.Materials;
 using ISAAR.MSolve.Materials.VonMisesStress;
 using ISAAR.MSolve.Numerical.LinearAlgebra;
+using ISAAR.MSolve.Preprocessor.Meshes;
+using ISAAR.MSolve.Preprocessor.Meshes.Custom;
+using ISAAR.MSolve.Preprocessor.Meshes.GMSH;
 using ISAAR.MSolve.Problems;
 using ISAAR.MSolve.Solvers.Interfaces;
 using ISAAR.MSolve.Solvers.Skyline;
